+++ compression/lpc/pip_heap.pxd
@@ -1,3 +1,4 @@
+# cython: language_level=3, cdivision=True, boundscheck=False, wraparound=False, nonecheck=False, initializedcheck=False, infer_types=True
 ctypedef PIPNode* NodePtr
 ctypedef PIPNode** NodeDbPtr
 ctypedef PIPHeap* HeapPtr
@@ -6,9 +7,9 @@ cdef struct PIPNode:
     NodePtr left
     NodePtr right
     HeapPtr parent
-    int ts
-    int index
-    int order
+    Py_ssize_t ts
+    Py_ssize_t index
+    Py_ssize_t order
     double value
     double cache
 
@@ -16,13 +17,13 @@ cdef struct PIPHeap:
     NodePtr head
     NodePtr tail
     NodeDbPtr values
-    int size
-    int m_size
-    int global_order
+    Py_ssize_t size
+    Py_ssize_t m_size
+    Py_ssize_t global_order
 
 cdef double vertical_distance(NodePtr left, NodePtr node, NodePtr right)
 
-cdef void init_node(NodePtr node, HeapPtr heap, int i)
+cdef void init_node(NodePtr node, HeapPtr heap, Py_ssize_t i)
 
 cdef void update_cache(NodePtr node)
 
@@ -32,27 +33,27 @@ cdef PIPNode recycle(NodePtr node)
 
 cdef PIPNode clear(NodePtr node)
 
-cdef void init_heap(HeapPtr heap, int size)
+cdef void init_heap(HeapPtr heap, Py_ssize_t size)
 
-cdef NodePtr acquire_item(HeapPtr heap, int ts, double value)
+cdef NodePtr acquire_item(HeapPtr heap, Py_ssize_t ts, double value)
 
-cdef void add(HeapPtr heap, int ts, double value)
+cdef void add(HeapPtr heap, Py_ssize_t ts, double value)
 
-cdef  int notify_change(HeapPtr heap, int index)
+cdef  Py_ssize_t notify_change(HeapPtr heap, Py_ssize_t index)
 
-cdef  int min(HeapPtr heap, int i, int j, int k)
+cdef  Py_ssize_t min(HeapPtr heap, Py_ssize_t i, Py_ssize_t j, Py_ssize_t k)
 
-cdef PIPNode remove_at(HeapPtr heap,  int index)
+cdef PIPNode remove_at(HeapPtr heap,  Py_ssize_t index)
 
-cdef  int bubble_up(HeapPtr heap,  int n)
+cdef  Py_ssize_t bubble_up(HeapPtr heap,  Py_ssize_t n)
 
-cdef  int bubble_down(HeapPtr heap,  int n)
+cdef  Py_ssize_t bubble_down(HeapPtr heap,  Py_ssize_t n)
 
-cdef  int swap(HeapPtr heap,  int i,  int j)
+cdef  Py_ssize_t swap(HeapPtr heap,  Py_ssize_t i,  Py_ssize_t j)
 
-cdef bint less(HeapPtr heap,  int i,  int j)
+cdef bint less(HeapPtr heap,  Py_ssize_t i,  Py_ssize_t j)
 
-cdef bint i_smaller_than_j(HeapPtr heap,  int i,  int j)
+cdef bint i_smaller_than_j(HeapPtr heap,  Py_ssize_t i,  Py_ssize_t j)
 
 cdef void iterate(HeapPtr heap)
